backend/requirements.txt
Flask
Flask-PyMongo
Flask-CORS
Flask-JWT-Extended
python-dotenv
bcrypt
email-validator
pymongo
dnspython
certifi